+++ src/main/java/csx55/domain/PredecessorNode.java
@@ -4,7 +4,7 @@ import java.io.Serializable;
 
 public class PredecessorNode extends ChordNode implements Serializable {
     private static final long serialversionUID = 1L;
-    public PredecessorNode(String descriptor, int peerId) {
+    public PredecessorNode(String descriptor, long peerId) {
         super(descriptor, peerId);
     }
 }
