+++ src/main/java/csx55/domain/SuccessorNode.java
@@ -5,7 +5,7 @@ import java.io.Serializable;
 public class SuccessorNode extends ChordNode implements Serializable  {
     private static final long serialversionUID = 1L;
 
-    public SuccessorNode(String descriptor, int peerId) {
+    public SuccessorNode(String descriptor, long peerId) {
         super(descriptor, peerId);
     }
 }
